refactor: reorganize workflow services and remove unused code
- Updated the import path for executeStreamingWorkflow to point to the correct services directory.
- Moved shared components and functions to a new sharedLangGraphComponents.ts file for better organization.
- Removed unused types and functions related to workflow steps, including WorkflowStepResult and its associated type guards.
- Streamlined the workflow services by consolidating common logic and ensuring clarity in the code structure.

This refactor enhances maintainability and readability of the workflow services.
