+++ frontend/internal-packages/agent/src/chat/workflow/index.ts
@@ -1,5 +1,5 @@
 import { LangGraphWorkflow } from './services/nonStreamingWorkflow'
-import { executeStreamingWorkflow } from './streaming/streamingWorkflow'
+import { executeStreamingWorkflow } from './services/streamingWorkflow'
 import type { ResponseChunk, WorkflowOptions, WorkflowState } from './types'
 
 /**
